+++ R/file.R
@@ -1,28 +1,34 @@
-#' @title check integrity of gzip files in a directory
+#' @title check integrity of `.gz` files in a directory
 #'
-#' @param path character. path to the folder(s) containing gzip files
+#' @param path character. path to the folder(s) containing `.gz` files.
 #'
-#' @return character. path of broken gzip files
+#' @return character. path of broken `.gz` files.
+#'
+#' @examples
+#' dir(system.file('extdata', package = 'libzhuoer'), full = T)
+#'
+#' check_gzip(system.file('extdata', package = 'libzhuoer'))
+#'
 #' @export
-#'
-#' @examples NULL
 check_gzip <- function(path) {
 	files <- dir(path, '\\.gz$', full.names = T);
 	okay <- parallel::mclapply(files, gzip_integrity) %>% sapply(as.logical);
 	files[!okay];
 }
 
-#' @title test gzip compressed file integrity
+#' @title check `.gz` file integrity
+#'
+#' @param path character. path to the `.gz` file to be checked.
+#'
+#' @return logical scalar. whether the `.gz` file is integrated.
+#'
+#' @examples
+#' gzip_integrity(system.file('extdata/good.gz', package = 'libzhuoer'))
+#' gzip_integrity(system.file('extdata/bad.gz', package = 'libzhuoer'))
+#'
+#' @export
+# path = 'inst/extdata/bad.gz'
 gzip_integrity <- function(path) {
-	system2('gzip', c('-t', path), T) %>% attr("status") %>% identical(NULL);
-}
-
-
-#' create a new directory and make sure it's empty
-#' export
-initiate_dir <- function(path) {
-    if (dir.exists(path))
-    	unlink(path, recursive = T)
-	else
-    	dir.create(path, recursive = T);
+	suppressWarnings( system2('gzip', c('-t', path), T, F) ) %>%
+		attr("status") %>% identical(NULL)
 }