+++ R/html.R
@@ -1,6 +1,6 @@
 #' @title open multiple urls sequentially
 #'
-#' @details the sequence maybe unstable when file number become large
+#' @details The sequence maybe unstable when file number become large
 #'
 #' @details It depends on external command (`system('sleep 0.2')`).
 #'
@@ -10,10 +10,13 @@
 #'   same order as that of `urls`
 #'
 #' @return `NULL`
-#' @export
 #'
-#' @section demo: `libzhuoer::browse_url(c('https://www.bing.com/',
-#'   'https://zhuoer.netlify.com/'), 0.4)`
+#' @examples
+#' \dontrun{
+#'     browse_url(c('https://www.bing.com/', 'https://dongzhuoer.com/'), 0.4)
+#' }
+#'
+#' @export
 browse_url <- function(urls, t = 0.2) {
 	plyr::l_ply(urls, . %>% {browseURL(.); system(paste('sleep', t))});