+++ R/main.R
@@ -10,7 +10,12 @@
 #' @return logical scalar
 #' @export
 #'
-#' @examples print_or_T(letters)
+#' @examples
+#' print_or_T(logical())
+#' print_or_T(letters)
+#'
+#' print_or_T(tibble::tibble())
+#' print_or_T(tibble::tibble(x = 1:5))
 print_or_T <- function(x) {
 	if (identical(length(x), 0L) || identical(nrow(x), 0L))
 		T
@@ -22,20 +27,11 @@ print_or_T <- function(x) {
 
 
 
-
-
-
-
-
-
-
-
-
 #' @title get nested elements
 #'
 #' @description get an element (usually a list) itself and elments of it, elements of elments of it, ...
 #'
-#' @details though `get_element` restrict the type of elemnt (e.g. the default value force element must be a list), the outermost elment isn't affected. The elements of `nested_element(list(1))` are all granted to be lists, but the first (and only) elment `nested_element(1)` is numeric, NOT list.
+#' @details Although `get_element` restrict the type of elemnt (e.g. the default value force element must be a list), the outermost elment isn't affected. The elements of `nested_element(list(1))` are all granted to be lists, but the first (and only) elment `nested_element(1)` is numeric, NOT list.
 #'
 #' @param element usually list. the outermost element.
 #' @param get_element function. recieve an element and return a list of element.